CRUD operations and authentication

+++ controllers/contact.controllers.js
@@ -1,17 +1,11 @@
 const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contact.model");
 
-//@description Get all contacts
-//@route GET /api/contacts
-//@access public
 const getAllContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
   res.status(200).json(contacts);
 });
 
-//@description create contacts
-//@route POST /api/contacts
-//@access public
 const createContact = asyncHandler(async (req, res) => {
   const { name, email, phone } = req.body;
 
@@ -23,13 +17,11 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
   res.status(201).json(contact);
 });
 
-//@description Get a particular contacts
-//@route GET /api/contacts/:id
-//@access public
 const getContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
@@ -39,15 +31,19 @@ const getContact = asyncHandler(async (req, res) => {
   res.status(200).json(contact);
 });
 
-//@description update contact
-//@route PUT /api/contacts/:id
-//@access public
 const updateContacts = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
 
   if (!contact) {
     res.status(404);
     throw new Error("Contact not found");
+  }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error(
+      "User don't have permission to update other user's contacts"
+    );
   }
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
@@ -57,17 +53,20 @@ const updateContacts = asyncHandler(async (req, res) => {
   res.status(200).json(updatedContact);
 });
 
-//@description delete a contact
-//@route DELETE /api/contacts/:id
-//@access public
 const deleteContacts = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Contact not found");
   }
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error(
+      "User don't have permission to update other user's contacts"
+    );
+  }
 
-  await Contact.deleteOne();
+  await Contact.deleteOne({ user_id: req.params.id });
   res.status(200).json(contact);
 });
